+++ co_cli/daemons/dream/_state.py
@@ -46,6 +46,7 @@ class HousekeepingStats(BaseModel):
     memory_decayed: int = 0
     skill_merged: int = 0
     skill_decayed: int = 0
+    done_pruned: int = 0
 
 
 class HousekeepingState(BaseModel):